Add metaballs basic code with only two blobs

+++ eggv2/utils.h
@@ -7,6 +7,7 @@ float distance(float x1, float y1, float x2, float y2) {
   return (float)sqrt(pow((x1 - x2), 2) + pow((y1 - y2), 2));
 }
 
-bool isBetween(int value, int min, int max, bool inclusive = false) {
+bool isBetween(int16_t value, int16_t min, int16_t max,
+               bool inclusive = false) {
   return inclusive ? value >= min && value <= max : value > min && value < max;
 }
